requirements.txt
discord.py==1.3.1
asyncpg==0.20.1
patreon
